chore: properly sanitizes admin components from client config
