+++ packages/dev/src/collections/Pages/CustomView.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const CustomView: React.FC = async (props) => {
+  return <div>CustomView</div>
+}
