+++ packages/ui/tsconfig.json
@@ -28,7 +28,7 @@
     "src/**/*.tsx",
     "src/**/*.d.ts",
     "src/**/*.json",
-    "../next/src/createClientConfig.ts"
+    "../next/src/utilities/createClientConfig.ts"
   ],
   "references": [{ "path": "../payload" }] // db-postgres depends on payload
 }
